refactor(backend): move shared response contracts and the blog feed into the domain slices
The MCP tools will soon move to their own OAuth-protected host. For a second
host to serve the same domain without referencing the web host, the pieces both
front doors share must live in libraries, not in Kalandra.Api.

- Response contracts served by both REST and MCP (GetJobOfferDetailResponse,
  CommentResponse, BlogCommentResponse) move into Kalandra.{JobOffers,Blog}/Contracts/.
  They're already pure Serialize(entity) records with no ASP.NET coupling.
- The RSS-backed blog feed reader (BlogFeedClient, BlogFeedConfig, BlogPostSummary)
  moves into Kalandra.Blog/Feed/ behind an AddBlogFeed(config, env) registration.
- Request DTOs, API error enums, and REST-only response envelopes (pagination,
  history, stats) stay at the HTTP boundary in Kalandra.Api.

Behavior-preserving: the /mcp route still runs in-process and the wire shapes are
unchanged. Full backend suite green (229 tests). docs/backend-api.md and
docs/backend-architecture.md record the new contract-location rule.
