+++ backend/src/Kalandra.Api/Features/JobOffers/Contracts/ListJobOffersResponse.cs
@@ -1,4 +1,5 @@
 using Kalandra.Api.Infrastructure;
+using Kalandra.JobOffers.Contracts;
 
 namespace Kalandra.Api.Features.JobOffers.Contracts;
