+++ backend/src/Kalandra.Blog/Kalandra.Blog.csproj
@@ -10,6 +10,7 @@
     <PackageReference Include="Kalicz.StrongTypes" Version="1.1.1" />
     <PackageReference Include="Marten" Version="9.12.0" />
     <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="10.0.9" />
+    <PackageReference Include="Microsoft.Extensions.Http" Version="10.0.9" />
   </ItemGroup>
 
   <ItemGroup>
